refactor(openfoam): split the oversized sloshing modules by responsibility
Brings the modules this issue touches under the 400-line rule instead of
grandfathering them, which the plan required once the files were being edited.

The coupling module separates into data contracts, the interpolation model, and
the derived analyses. The derived analyses become free functions taking the
model, with thin methods left on the model so the supported entry point is
unchanged. Signatures were compared against the pre-split module and two
transcription defects in the delegating methods were caught and corrected that
way: a wrong default on the escalation threshold and a dropped optional on the
tuning entry point.

The 2D validation module separates into configuration contracts, case-dictionary
templates, case builders with their provenance, and post-processing. Because
many scripts and the validation package import from the original module path,
that path becomes a facade re-exporting every name it previously exposed, so no
caller changes. The sweep harness moves its first-harmonic reduction and its
command-line surface into siblings, keeping the reduction re-exported.

Function length is measured as logic lines with a leading docstring excluded, and
that basis is stated in the test rather than left implicit. Every function in the
governed set was already inside the limit on that basis, so none was split to
satisfy a line count.

Full openfoam suite: 723 passed, 3 failed. The three failures are the recorded
pre-existing numpy failures and are byte-identical to the baseline node IDs.

Refs #1574
